services/simulation/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.0
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
python-dotenv==1.0.0
httpx==0.25.2
email-validator>=2.0.0
# MiroFish pipeline dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
zep-python>=2.0.0
